Big Update Drink Menu
Updated the view to read drinks from a text file. Drink is no longer an
interface, it is now a class. The drink class that implemented Drink
were deleted and will now just be object of type drink. The view now
creates drinks with ingredients by reading from a text file.

Next step is to generate the view based on the drinkMenu Array. Then
work on setting up action readers to submit user choices to the
controller.